+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # CHANGELOG.md
 
+## 0.1.26 (2023-02-26)
+
+- Updated `Form` components to show correct color based on form state.
+
+## 0.1.25 (2023-02-26)
+
+- Changed `defaultOption` prop to `defaultOptionId` in `FormSelect` component.
+
+## 0.1.24 (2023-02-26)
+
+- Added `defaultOption` prop to `FormSelect` component.
+
+## 0.1.23 (2023-02-25)
+
+- Added function to calculate page numbers for `Pagination` component.
+
+## 0.1.22 (2023-02-24)
+
+- Fixed default value not appearing in `TypeaheadSelect` component.
+
 ## 0.1.20 (2023-02-15)
 
 - Added ref to `Modal` component.